Create README.md
Hi, you can contribute to this checklist and add your own question to the repository. When there are more questions, I'm going to make a surprise mobile app.... :-) Please avoid any use of bad language. The checklist is supposed to be from life, as well as fun, as well as enriching. Thanks!